♻️ [Refactoring]: [BE] task ドメインを 13 ファイルに分割し DDD 戦術的パターンへ再編
旧 task/index.rs (950L) を task/ 直下 7 ファイル (warning / task_index / parse /
parent_validation / children / reverse_links / path_lookup) にフラット分割し、
旧 task/create.rs (564L) を create/ 配下 6 ファイル (args / error / filename /
content / command + 親 create.rs ファサード) に分割した。

- validate.rs は作らず、validation を TaskIndex aggregate のメソッド
  (validate_new_parent / validate_with_new_task) に集約
- 新規 VO TaskContent (scanner eligible を constructor で強制) を追加
- TaskFileName::from_title を build_new_filename から委譲利用
- TaskExtras 型エイリアス廃止、BTreeMap を直接利用
- parent_validation / children / reverse_links / path_lookup の自由関数は
  pub(super) に格下げし、task ドメイン外からは TaskIndex 経由で利用
- 外部 import 7 ファイル (state / project / watcher_event / task::get) を新パスへ書き換え
- docs/impl/ddd-domain-design.md に "9. task ドメインの内部分割" 節を追記
- CLAUDE.md のディレクトリ構成セクションを新構造に更新

公開振る舞い (Tauri command I/F / IPC payload / エラー文字列 / warning code) は完全互換。
既存 107 テスト (index 75 + create 32) は維持し、新規追加は TaskContent VO の境界
テスト 5 件のみ。
