--- src-tauri/src/task/index_tests.rs
@@ -1,1315 +1,0 @@
-use super::super::frontmatter::{Frontmatter, Priority};
-use super::*;
-use serde_json::json;
-use std::collections::BTreeMap;
-
-fn context(path: &str) -> TaskParseContext {
-    TaskParseContext {
-        file_path: PathBuf::from(path),
-        default_status: "Todo".into(),
-    }
-}
-
-fn task_from(input: &str, path: &str) -> Task {
-    task_from_markdown(input.as_bytes(), &context(path)).unwrap()
-}
-
-fn task_with_parent(path: &str, parent: &str) -> Task {
-    task_from(
-        &format!("---\ntitle: Task\nstatus: Todo\nparent: {parent}\n---\n"),
-        path,
-    )
-}
-
-fn task_without_parent(path: &str) -> Task {
-    task_from("---\ntitle: Task\nstatus: Todo\n---\n", path)
-}
-
-fn task_with_links(path: &str, links: &[&str]) -> Task {
-    let links_yaml = links
-        .iter()
-        .map(|link| format!("  - {link}"))
-        .collect::<Vec<_>>()
-        .join("\n");
-    task_from(
-        &format!("---\ntitle: Task\nstatus: Todo\nlinks:\n{links_yaml}\n---\n"),
-        path,
-    )
-}
-
-fn parent_chain_with_edge_count(edge_count: usize) -> Vec<Task> {
-    let mut tasks = Vec::new();
-
-    for index in 0..edge_count {
-        tasks.push(task_with_parent(
-            &format!("tasks/{index}.md"),
-            &format!("tasks/{}.md", index + 1),
-        ));
-    }
-
-    tasks.push(task_without_parent(&format!("tasks/{edge_count}.md")));
-    tasks
-}
-
-fn parsed_with_extras(extras: serde_yaml_ng::Mapping) -> Parsed {
-    Parsed {
-        frontmatter: Frontmatter {
-            extras,
-            ..Frontmatter::default()
-        },
-        body: String::new(),
-    }
-}
-
-#[test]
-fn minimum_frontmatter_generates_task() {
-    let task = task_from(
-        "---\ntitle: Fix bug\nstatus: Doing\n---\n",
-        "tasks/fix-bug.md",
-    );
-
-    assert_eq!(task.id, "tasks/fix-bug.md");
-    assert_eq!(task.file_path, "tasks/fix-bug.md");
-    assert_eq!(task.title, "Fix bug");
-    assert_eq!(task.status, "Doing");
-    assert_eq!(task.priority, None);
-    assert_eq!(task.labels, Vec::<Label>::new());
-    assert_eq!(task.parent, None);
-    assert_eq!(task.links, Vec::<TaskFilePath>::new());
-    assert_eq!(task.children, Vec::<TaskFilePath>::new());
-    assert_eq!(task.reverse_links, Vec::<TaskFilePath>::new());
-    assert_eq!(task.body, "");
-    assert_eq!(task.extras, BTreeMap::new());
-    assert_eq!(task.warnings, Vec::<TaskWarning>::new());
-}
-
-#[test]
-fn typed_parser_fields_and_body_are_reflected() {
-    let task = task_from(
-        "---\ntitle: Feature\nstatus: Doing\npriority: high\nlabels: [bug, bug, api]\nlinks: related.md\n---\nBody\n",
-        "feature.md",
-    );
-
-    assert_eq!(task.priority, Some(Priority::High));
-    assert_eq!(task.labels, vec![Label::from("bug"), Label::from("api")]);
-    assert_eq!(task.links, vec![TaskFilePath::from("related.md")]);
-    assert_eq!(task.body, "Body\n");
-}
-
-#[test]
-fn missing_title_uses_file_name_fallback_with_warning() {
-    let task = task_from("---\nstatus: Todo\n---\n", "tasks/fix-login.md");
-
-    assert_eq!(task.title, "fix login");
-    assert_eq!(
-        task.warnings,
-        vec![TaskWarning {
-            code: TaskWarningCode::MissingTitleUsedFileName,
-            field: Some("title".into()),
-            message: "title is missing; file name was used".to_string(),
-        }]
-    );
-}
-
-#[test]
-fn invalid_title_uses_file_name_fallback_with_warning() {
-    let cases = [
-        ("---\ntitle: ''\nstatus: Todo\n---\n", "空文字"),
-        ("---\ntitle: 123\nstatus: Todo\n---\n", "非文字列"),
-    ];
-
-    for (input, label) in cases {
-        let task = task_from(input, "tasks/fix-login.md");
-        assert_eq!(task.title, "fix login", "{label}");
-        assert_eq!(
-            task.warnings,
-            vec![TaskWarning {
-                code: TaskWarningCode::InvalidTitleUsedFileName,
-                field: Some("title".into()),
-                message: "title is invalid; file name was used".to_string(),
-            }],
-            "{label}"
-        );
-    }
-}
-
-#[test]
-fn missing_status_uses_default_with_warning() {
-    let task = task_from("---\ntitle: Fix bug\n---\n", "tasks/fix-bug.md");
-
-    assert_eq!(task.status, "Todo");
-    assert_eq!(
-        task.warnings,
-        vec![TaskWarning {
-            code: TaskWarningCode::MissingStatusUsedDefault,
-            field: Some("status".into()),
-            message: "status is missing; default status was used".to_string(),
-        }]
-    );
-}
-
-#[test]
-fn invalid_status_uses_default_with_warning() {
-    let task = task_from(
-        "---\ntitle: Fix bug\nstatus: [Doing]\n---\n",
-        "tasks/fix-bug.md",
-    );
-
-    assert_eq!(task.status, "Todo");
-    assert_eq!(
-        task.warnings,
-        vec![TaskWarning {
-            code: TaskWarningCode::InvalidStatusUsedDefault,
-            field: Some("status".into()),
-            message: "status is invalid; default status was used".to_string(),
-        }]
-    );
-}
-
-#[test]
-fn parent_is_reflected_when_string_and_ignored_when_missing_or_invalid() {
-    let parent_task = task_from(
-        "---\ntitle: Child\nstatus: Todo\nparent: tasks/parent.md\n---\n",
-        "tasks/child.md",
-    );
-    let missing_parent_task = task_from("---\ntitle: Root\nstatus: Todo\n---\n", "tasks/root.md");
-    let invalid_parent_task = task_from(
-        "---\ntitle: Root\nstatus: Todo\nparent: 123\n---\n",
-        "tasks/root.md",
-    );
-
-    assert_eq!(parent_task.parent, Some("tasks/parent.md".into()));
-    assert_eq!(missing_parent_task.parent, None);
-    assert_eq!(invalid_parent_task.parent, None);
-    assert_eq!(
-        invalid_parent_task.warnings,
-        vec![TaskWarning {
-            code: TaskWarningCode::InvalidParentIgnored,
-            field: Some("parent".into()),
-            message: "parent is invalid; value was ignored".to_string(),
-        }]
-    );
-}
-
-#[test]
-fn parent_existence_validation_does_not_warn_when_parent_is_missing_or_existing() {
-    let tasks = validate_parent_existence(vec![
-        task_from("---\ntitle: Root\nstatus: Todo\n---\n", "tasks/root.md"),
-        task_from("---\ntitle: Parent\nstatus: Todo\n---\n", "tasks/parent.md"),
-        task_from(
-            "---\ntitle: Child\nstatus: Todo\nparent: tasks/parent.md\n---\n",
-            "tasks/child.md",
-        ),
-    ]);
-
-    assert!(tasks[0].warnings.is_empty());
-    assert!(tasks[2]
-        .warnings
-        .iter()
-        .all(|warning| warning.code != TaskWarningCode::ParentNotFound));
-}
-
-#[test]
-fn missing_parent_adds_warning_without_dropping_task_or_parent_value() {
-    let tasks = validate_parent_existence(vec![task_from(
-        "---\ntitle: Child\nstatus: Todo\nparent: tasks/missing.md\n---\n",
-        "tasks/child.md",
-    )]);
-
-    assert_eq!(tasks[0].parent, Some("tasks/missing.md".into()));
-    assert!(tasks[0].warnings.iter().any(|warning| {
-        warning.code == TaskWarningCode::ParentNotFound
-            && warning.field.as_deref() == Some("parent")
-    }));
-}
-
-#[test]
-fn empty_parent_adds_warning_without_normalizing_parent_value() {
-    let tasks = validate_parent_existence(vec![task_from(
-        "---\ntitle: Child\nstatus: Todo\nparent: ''\n---\n",
-        "tasks/child.md",
-    )]);
-
-    assert_eq!(tasks[0].parent, Some(String::new().into()));
-    assert!(tasks[0].warnings.iter().any(|warning| {
-        warning.code == TaskWarningCode::ParentNotFound
-            && warning.field.as_deref() == Some("parent")
-    }));
-}
-
-#[test]
-fn parent_existence_validation_warns_each_task_and_keeps_existing_warnings() {
-    let tasks = validate_parent_existence(vec![
-        task_from(
-            "---\ntitle: Child\nstatus: Todo\nparent: tasks/missing-a.md\n---\n",
-            "tasks/child-a.md",
-        ),
-        task_from(
-            "---\ntitle: 123\nstatus: Todo\nparent: tasks/missing-b.md\n---\n",
-            "tasks/child-b.md",
-        ),
-    ]);
-
-    assert!(tasks[0].warnings.iter().any(|warning| {
-        warning.code == TaskWarningCode::ParentNotFound
-            && warning.field.as_deref() == Some("parent")
-    }));
-    assert!(tasks[1]
-        .warnings
-        .iter()
-        .any(|warning| warning.code == TaskWarningCode::InvalidTitleUsedFileName));
-    assert!(tasks[1].warnings.iter().any(|warning| {
-        warning.code == TaskWarningCode::ParentNotFound
-            && warning.field.as_deref() == Some("parent")
-    }));
-}
-
-#[test]
-fn parent_existence_validation_does_not_duplicate_parent_not_found_warning() {
-    let mut task = task_from(
-        "---\ntitle: Child\nstatus: Todo\nparent: tasks/missing.md\n---\n",
-        "tasks/child.md",
-    );
-    task.warnings.push(TaskWarning {
-        code: TaskWarningCode::ParentNotFound,
-        field: Some("parent".into()),
-        message: "parent task was not found".to_string(),
-    });
-
-    let tasks = validate_parent_existence(vec![task]);
-    let warning_count = tasks[0]
-        .warnings
-        .iter()
-        .filter(|warning| {
-            warning.code == TaskWarningCode::ParentNotFound
-                && warning.field.as_deref() == Some("parent")
-        })
-        .count();
-
-    assert_eq!(warning_count, 1);
-}
-
-#[test]
-fn self_parent_is_treated_as_existing_parent() {
-    let tasks = validate_parent_existence(vec![task_from(
-        "---\ntitle: Child\nstatus: Todo\nparent: tasks/child.md\n---\n",
-        "tasks/child.md",
-    )]);
-
-    assert!(tasks[0]
-        .warnings
-        .iter()
-        .all(|warning| warning.code != TaskWarningCode::ParentNotFound));
-}
-
-#[test]
-fn build_children_adds_child_file_path_to_parent() {
-    let tasks = vec![
-        task_without_parent("tasks/parent.md"),
-        task_with_parent("tasks/child.md", "tasks/parent.md"),
-    ];
-
-    let tasks = build_children(tasks).unwrap();
-
-    assert_eq!(
-        tasks[0].children,
-        vec![TaskFilePath::from("tasks/child.md")]
-    );
-}
-
-#[test]
-fn build_children_adds_child_file_paths_to_parent_in_input_order() {
-    let tasks = vec![
-        task_without_parent("tasks/parent.md"),
-        task_with_parent("tasks/child-a.md", "tasks/parent.md"),
-        task_with_parent("tasks/child-b.md", "tasks/parent.md"),
-    ];
-
-    let tasks = build_children(tasks).unwrap();
-
-    assert_eq!(
-        tasks[0].children,
-        vec![
-            "tasks/child-a.md".to_string(),
-            "tasks/child-b.md".to_string(),
-        ]
-    );
-}
-
-#[test]
-fn build_children_adds_child_when_parent_appears_later() {
-    let tasks = vec![
-        task_with_parent("tasks/child.md", "tasks/parent.md"),
-        task_without_parent("tasks/parent.md"),
-    ];
-
-    let tasks = build_children(tasks).unwrap();
-
-    assert_eq!(
-        tasks[1].children,
-        vec![TaskFilePath::from("tasks/child.md")]
-    );
-}
-
-#[test]
-fn build_children_clears_existing_children_before_recalculation() {
-    let mut parent = task_without_parent("tasks/parent.md");
-    parent.children = vec![
-        "tasks/stale.md".into(),
-        "tasks/child.md".into(),
-        "tasks/child.md".into(),
-    ];
-    let tasks = vec![
-        parent,
-        task_with_parent("tasks/child.md", "tasks/parent.md"),
-    ];
-
-    let tasks = build_children(tasks).unwrap();
-
-    assert_eq!(
-        tasks[0].children,
-        vec![TaskFilePath::from("tasks/child.md")]
-    );
-}
-
-#[test]
-fn build_children_matches_parent_with_dot_prefix() {
-    let tasks = vec![
-        task_without_parent("tasks/parent.md"),
-        task_with_parent("tasks/child.md", "./tasks/parent.md"),
-    ];
-
-    let tasks = build_children(tasks).unwrap();
-
-    assert_eq!(
-        tasks[0].children,
-        vec![TaskFilePath::from("tasks/child.md")]
-    );
-}
-
-#[test]
-fn build_children_matches_parent_with_backslash_separator() {
-    let tasks = vec![
-        task_without_parent("tasks/parent.md"),
-        task_with_parent("tasks/child.md", "tasks\\parent.md"),
-    ];
-
-    let tasks = build_children(tasks).unwrap();
-
-    assert_eq!(
-        tasks[0].children,
-        vec![TaskFilePath::from("tasks/child.md")]
-    );
-}
-
-#[test]
-fn build_children_keeps_missing_parent_warning_without_child_append() {
-    let tasks = vec![
-        task_without_parent("tasks/parent.md"),
-        task_with_parent("tasks/child.md", "tasks/missing.md"),
-    ];
-
-    let tasks = build_children(tasks).unwrap();
-
-    assert!(tasks[0].children.is_empty());
-    assert!(tasks[1].warnings.iter().any(|warning| {
-        warning.code == TaskWarningCode::ParentNotFound
-            && warning.field.as_deref() == Some("parent")
-    }));
-}
-
-#[test]
-fn build_children_ignores_empty_absolute_and_drive_prefix_parent_for_child_append() {
-    let cases = ["", "/tasks/parent.md", "C:\\tasks\\parent.md"];
-
-    for parent in cases {
-        let mut child = task_without_parent("tasks/child.md");
-        child.parent = Some(parent.into());
-        let tasks = vec![task_without_parent("tasks/parent.md"), child];
-
-        let tasks = build_children(tasks).unwrap();
-
-        assert!(tasks[0].children.is_empty(), "{parent}");
-        assert!(
-            tasks[1].warnings.iter().any(|warning| {
-                warning.code == TaskWarningCode::ParentNotFound
-                    && warning.field.as_deref() == Some("parent")
-            }),
-            "{parent}"
-        );
-    }
-}
-
-#[test]
-fn build_children_returns_error_for_self_reference() {
-    let result = build_children(vec![task_with_parent("tasks/a.md", "tasks/a.md")]);
-
-    assert!(matches!(
-        result,
-        Err(TaskParseError::CycleOrTooDeep {
-            file_path,
-            reason: ParentHierarchyErrorReason::Cycle,
-        }) if file_path == "tasks/a.md"
-    ));
-}
-
-#[test]
-fn build_children_returns_error_for_cycle() {
-    let result = build_children(vec![
-        task_with_parent("tasks/a.md", "tasks/b.md"),
-        task_with_parent("tasks/b.md", "tasks/a.md"),
-    ]);
-
-    assert!(matches!(
-        result,
-        Err(TaskParseError::CycleOrTooDeep {
-            file_path,
-            reason: ParentHierarchyErrorReason::Cycle,
-        }) if file_path == "tasks/a.md"
-    ));
-}
-
-#[test]
-fn build_children_returns_error_for_parent_chain_over_max_depth() {
-    let result = build_children(parent_chain_with_edge_count(21));
-
-    assert!(matches!(
-        result,
-        Err(TaskParseError::CycleOrTooDeep {
-            file_path,
-            reason: ParentHierarchyErrorReason::TooDeep,
-        }) if file_path == "tasks/0.md"
-    ));
-}
-
-#[test]
-fn build_children_accepts_empty_tasks() {
-    let tasks = build_children(Vec::new()).unwrap();
-
-    assert!(tasks.is_empty());
-}
-
-#[test]
-fn build_reverse_links_adds_source_file_path_to_target() {
-    let tasks = vec![
-        task_with_links("tasks/source.md", &["tasks/target.md"]),
-        task_without_parent("tasks/target.md"),
-    ];
-
-    let tasks = build_reverse_links(tasks);
-
-    assert_eq!(
-        tasks[1].reverse_links,
-        vec![TaskFilePath::from("tasks/source.md")]
-    );
-}
-
-#[test]
-fn build_reverse_links_adds_sources_in_input_order() {
-    let tasks = vec![
-        task_with_links("tasks/source-a.md", &["tasks/target.md"]),
-        task_with_links("tasks/source-b.md", &["tasks/target.md"]),
-        task_without_parent("tasks/target.md"),
-    ];
-
-    let tasks = build_reverse_links(tasks);
-
-    assert_eq!(
-        tasks[2].reverse_links,
-        vec![
-            "tasks/source-a.md".to_string(),
-            "tasks/source-b.md".to_string(),
-        ]
-    );
-}
-
-#[test]
-fn build_reverse_links_adds_source_to_each_existing_link_target() {
-    let tasks = vec![
-        task_with_links(
-            "tasks/source.md",
-            &["tasks/target-b.md", "tasks/target-a.md"],
-        ),
-        task_without_parent("tasks/target-a.md"),
-        task_without_parent("tasks/target-b.md"),
-    ];
-
-    let tasks = build_reverse_links(tasks);
-
-    assert_eq!(
-        tasks[1].reverse_links,
-        vec![TaskFilePath::from("tasks/source.md")]
-    );
-    assert_eq!(
-        tasks[2].reverse_links,
-        vec![TaskFilePath::from("tasks/source.md")]
-    );
-}
-
-#[test]
-fn build_reverse_links_deduplicates_normalized_targets_per_source() {
-    let tasks = vec![
-        task_with_links(
-            "tasks/source.md",
-            &["tasks/target.md", "./tasks/target.md", "tasks\\target.md"],
-        ),
-        task_without_parent("tasks/target.md"),
-    ];
-
-    let tasks = build_reverse_links(tasks);
-
-    assert_eq!(
-        tasks[1].reverse_links,
-        vec![TaskFilePath::from("tasks/source.md")]
-    );
-}
-
-#[test]
-fn build_reverse_links_clears_existing_reverse_links_before_recalculation() {
-    let mut target = task_without_parent("tasks/target.md");
-    target.reverse_links = vec![
-        "tasks/stale.md".into(),
-        "tasks/source.md".into(),
-        "tasks/source.md".into(),
-    ];
-    let tasks = vec![
-        task_with_links("tasks/source.md", &["tasks/target.md"]),
-        target,
-    ];
-
-    let tasks = build_reverse_links(tasks);
-
-    assert_eq!(
-        tasks[1].reverse_links,
-        vec![TaskFilePath::from("tasks/source.md")]
-    );
-}
-
-#[test]
-fn build_reverse_links_ignores_missing_target() {
-    let tasks = vec![task_with_links("tasks/source.md", &["tasks/missing.md"])];
-
-    let tasks = build_reverse_links(tasks);
-
-    assert!(tasks[0].reverse_links.is_empty());
-    assert!(tasks[0].warnings.is_empty());
-}
-
-#[test]
-fn build_reverse_links_matches_link_with_dot_prefix() {
-    let tasks = vec![
-        task_with_links("tasks/source.md", &["./tasks/target.md"]),
-        task_without_parent("tasks/target.md"),
-    ];
-
-    let tasks = build_reverse_links(tasks);
-
-    assert_eq!(
-        tasks[1].reverse_links,
-        vec![TaskFilePath::from("tasks/source.md")]
-    );
-}
-
-#[test]
-fn build_reverse_links_matches_link_with_backslash_separator() {
-    let tasks = vec![
-        task_with_links("tasks/source.md", &["tasks\\target.md"]),
-        task_without_parent("tasks/target.md"),
-    ];
-
-    let tasks = build_reverse_links(tasks);
-
-    assert_eq!(
-        tasks[1].reverse_links,
-        vec![TaskFilePath::from("tasks/source.md")]
-    );
-}
-
-#[test]
-fn build_reverse_links_ignores_empty_absolute_and_drive_prefix_link() {
-    let cases = [
-        "",
-        "/tasks/target.md",
-        "\\tasks\\target.md",
-        "C:\\tasks\\target.md",
-    ];
-
-    for link in cases {
-        let mut source = task_without_parent("tasks/source.md");
-        source.links = vec![link.into()];
-        let tasks = vec![source, task_without_parent("tasks/target.md")];
-
-        let tasks = build_reverse_links(tasks);
-
-        assert!(tasks[1].reverse_links.is_empty(), "{link}");
-    }
-}
-
-#[test]
-fn build_reverse_links_accepts_empty_tasks() {
-    let tasks = build_reverse_links(Vec::new());
-
-    assert!(tasks.is_empty());
-}
-
-#[test]
-fn build_reverse_links_allows_self_link() {
-    let tasks = vec![task_with_links("tasks/source.md", &["tasks/source.md"])];
-
-    let tasks = build_reverse_links(tasks);
-
-    assert_eq!(
-        tasks[0].reverse_links,
-        vec![TaskFilePath::from("tasks/source.md")]
-    );
-}
-
-#[test]
-fn direct_cycle_returns_cycle_or_too_deep() {
-    let result = validate_parent_hierarchy(vec![task_with_parent("tasks/a.md", "tasks/a.md")]);
-
-    assert!(matches!(
-        result,
-        Err(TaskParseError::CycleOrTooDeep {
-            file_path,
-            reason: ParentHierarchyErrorReason::Cycle,
-        }) if file_path == "tasks/a.md"
-    ));
-}
-
-#[test]
-fn multi_node_cycle_returns_cycle_or_too_deep() {
-    let result = validate_parent_hierarchy(vec![
-        task_with_parent("tasks/a.md", "tasks/b.md"),
-        task_with_parent("tasks/b.md", "tasks/c.md"),
-        task_with_parent("tasks/c.md", "tasks/a.md"),
-    ]);
-
-    assert!(matches!(
-        result,
-        Err(TaskParseError::CycleOrTooDeep {
-            file_path,
-            reason: ParentHierarchyErrorReason::Cycle,
-        }) if file_path == "tasks/a.md"
-    ));
-}
-
-#[test]
-fn depth_over_20_returns_cycle_or_too_deep() {
-    let result = validate_parent_hierarchy(parent_chain_with_edge_count(21));
-
-    assert!(matches!(
-        result,
-        Err(TaskParseError::CycleOrTooDeep {
-            file_path,
-            reason: ParentHierarchyErrorReason::TooDeep,
-        }) if file_path == "tasks/0.md"
-    ));
-}
-
-#[test]
-fn depth_20_is_allowed() {
-    let result = validate_parent_hierarchy(parent_chain_with_edge_count(20));
-
-    assert!(result.is_ok());
-}
-
-#[test]
-fn missing_parent_keeps_warning_without_cycle_error() {
-    let tasks =
-        validate_parent_hierarchy(vec![task_with_parent("tasks/child.md", "tasks/missing.md")])
-            .unwrap();
-
-    assert_eq!(tasks[0].parent, Some("tasks/missing.md".into()));
-    assert!(tasks[0].warnings.iter().any(|warning| {
-        warning.code == TaskWarningCode::ParentNotFound
-            && warning.field.as_deref() == Some("parent")
-    }));
-}
-
-#[test]
-fn separator_variation_cycle_is_detected() {
-    let result = validate_parent_hierarchy(vec![
-        task_with_parent("tasks/a.md", ".\\tasks\\b.md"),
-        task_with_parent("tasks/b.md", "./tasks/a.md"),
-    ]);
-
-    assert!(matches!(
-        result,
-        Err(TaskParseError::CycleOrTooDeep {
-            file_path,
-            reason: ParentHierarchyErrorReason::Cycle,
-        }) if file_path == "tasks/a.md"
-    ));
-}
-
-#[test]
-fn cycle_or_too_deep_error_message_includes_file_path_and_reason() {
-    let result = validate_parent_hierarchy(vec![task_with_parent("tasks/a.md", "tasks/a.md")]);
-    let Err(error) = result else {
-        panic!("cycle should return error");
-    };
-
-    assert_eq!(
-        error.to_string(),
-        "parent chain for 'tasks/a.md' contains a cycle"
-    );
-}
-
-#[test]
-fn parent_lookup_accepts_separator_and_current_directory_variations() {
-    let cases = ["tasks\\parent.md", "./tasks/parent.md"];
-
-    for parent in cases {
-        let tasks = validate_parent_existence(vec![
-            task_from("---\ntitle: Parent\nstatus: Todo\n---\n", "tasks/parent.md"),
-            task_from(
-                &format!("---\ntitle: Child\nstatus: Todo\nparent: {parent}\n---\n"),
-                "tasks/child.md",
-            ),
-        ]);
-
-        assert!(
-            tasks[1]
-                .warnings
-                .iter()
-                .all(|warning| warning.code != TaskWarningCode::ParentNotFound),
-            "{parent}"
-        );
-    }
-}
-
-#[test]
-fn parent_lookup_rejects_absolute_or_drive_prefixed_parent_paths() {
-    let cases = ["/tasks/parent.md", "C:\\tasks\\parent.md"];
-
-    for parent in cases {
-        let tasks = validate_parent_existence(vec![
-            task_from("---\ntitle: Parent\nstatus: Todo\n---\n", "tasks/parent.md"),
-            task_from(
-                &format!("---\ntitle: Child\nstatus: Todo\nparent: {parent}\n---\n"),
-                "tasks/child.md",
-            ),
-        ]);
-
-        assert!(
-            tasks[1].warnings.iter().any(|warning| {
-                warning.code == TaskWarningCode::ParentNotFound
-                    && warning.field.as_deref() == Some("parent")
-            }),
-            "{parent}"
-        );
-    }
-}
-
-#[test]
-fn unknown_fields_are_kept_in_extras_as_json_values() {
-    let task = task_from(
-        "---\ntitle: Fix bug\nstatus: Todo\nestimate: 3\nmeta:\n  owner: alice\n---\n",
-        "tasks/fix-bug.md",
-    );
-
-    assert_eq!(task.extras.get("estimate"), Some(&json!(3)));
-    assert_eq!(task.extras.get("meta"), Some(&json!({ "owner": "alice" })));
-}
-
-#[test]
-fn non_string_extra_key_is_excluded_with_warning() {
-    let mut extras = serde_yaml_ng::Mapping::new();
-    extras.insert(
-        serde_yaml_ng::Value::String("title".to_string()),
-        serde_yaml_ng::Value::String("Fix bug".to_string()),
-    );
-    extras.insert(
-        serde_yaml_ng::Value::String("status".to_string()),
-        serde_yaml_ng::Value::String("Todo".to_string()),
-    );
-    extras.insert(
-        serde_yaml_ng::Value::Sequence(vec![serde_yaml_ng::Value::String("a".to_string())]),
-        serde_yaml_ng::Value::String("value".to_string()),
-    );
-    let task = task_from_parsed(parsed_with_extras(extras), &context("tasks/fix-bug.md"));
-
-    assert_eq!(task.extras, BTreeMap::new());
-    assert_eq!(
-        task.warnings,
-        vec![TaskWarning {
-            code: TaskWarningCode::NonStringExtraKeyIgnored,
-            field: None,
-            message: "non-string extra key was ignored".to_string(),
-        }]
-    );
-}
-
-#[test]
-fn json_incompatible_extra_value_is_excluded_with_warning() {
-    let mut extras = serde_yaml_ng::Mapping::new();
-    extras.insert(
-        serde_yaml_ng::Value::String("title".to_string()),
-        serde_yaml_ng::Value::String("Fix bug".to_string()),
-    );
-    extras.insert(
-        serde_yaml_ng::Value::String("status".to_string()),
-        serde_yaml_ng::Value::String("Todo".to_string()),
-    );
-    extras.insert(
-        serde_yaml_ng::Value::String("tagged".to_string()),
-        serde_yaml_ng::Value::Tagged(Box::new(serde_yaml_ng::value::TaggedValue {
-            tag: serde_yaml_ng::value::Tag::new("custom"),
-            value: serde_yaml_ng::Value::String("value".to_string()),
-        })),
-    );
-    let task = task_from_parsed(parsed_with_extras(extras), &context("tasks/fix-bug.md"));
-
-    assert_eq!(task.extras, BTreeMap::new());
-    assert_eq!(
-        task.warnings,
-        vec![TaskWarning {
-            code: TaskWarningCode::ExtraValueNotJsonCompatible,
-            field: Some("tagged".into()),
-            message: "extra value is not JSON compatible; value was ignored".to_string(),
-        }]
-    );
-}
-
-#[test]
-fn typed_keys_are_excluded_from_extras() {
-    let task = task_from(
-        "---\ntitle: Fix bug\nstatus: Todo\npriority: High\nlabels: [bug]\nparent: parent.md\nlinks: [related.md]\nextra: kept\n---\n",
-        "tasks/fix-bug.md",
-    );
-
-    assert_eq!(
-        task.extras,
-        BTreeMap::from([("extra".to_string(), json!("kept"))])
-    );
-}
-
-#[test]
-fn task_serializes_path_fields_and_warning_codes_as_camel_case() {
-    let task = task_from(
-        "---\nstatus: Todo\nestimate: 3\n---\nBody\n",
-        "tasks/fix-bug.md",
-    );
-
-    let json_value = serde_json::to_value(task).unwrap();
-
-    assert_eq!(json_value["filePath"], json!("tasks/fix-bug.md"));
-    assert_eq!(json_value["reverseLinks"], json!([]));
-    assert_eq!(json_value["extras"], json!({ "estimate": 3 }));
-    assert_eq!(
-        json_value["warnings"][0]["code"],
-        json!("missingTitleUsedFileName")
-    );
-}
-
-#[test]
-fn parent_not_found_warning_code_serializes_as_camel_case() {
-    let warning = TaskWarning {
-        code: TaskWarningCode::ParentNotFound,
-        field: Some("parent".into()),
-        message: "parent task was not found".to_string(),
-    };
-
-    let json_value = serde_json::to_value(warning).unwrap();
-
-    assert_eq!(json_value["code"], json!("parentNotFound"));
-}
-
-#[test]
-fn task_file_path_payload_is_relative_and_uses_forward_slashes() {
-    let cases = [
-        ("/project\\tasks\\fix-bug.md", "project/tasks/fix-bug.md"),
-        ("C:\\project\\tasks\\fix-bug.md", "project/tasks/fix-bug.md"),
-    ];
-
-    for (input_path, expected_path) in cases {
-        let task = task_from("---\ntitle: Fix bug\nstatus: Todo\n---\n", input_path);
-
-        assert_eq!(task.id, expected_path);
-        assert_eq!(task.file_path, expected_path);
-    }
-}
-
-#[test]
-fn frontmatter_absence_returns_not_task() {
-    let result = task_from_markdown(b"# Heading\nbody\n", &context("notes.md"));
-
-    assert!(matches!(result, Err(TaskParseError::NotTask)));
-}
-
-#[test]
-fn invalid_yaml_or_encoding_returns_frontmatter_error() {
-    let invalid_yaml = task_from_markdown(b"---\n: bad\n---\n", &context("bad.md"));
-    let invalid_encoding = task_from_markdown(b"\xff\xfe---\n---\n", &context("bad.md"));
-
-    assert!(matches!(invalid_yaml, Err(TaskParseError::Frontmatter(_))));
-    assert!(matches!(
-        invalid_encoding,
-        Err(TaskParseError::Frontmatter(_))
-    ));
-}
-
-#[test]
-fn resolve_parent_for_new_task_hit_cases() {
-    let tasks = vec![task_without_parent("tasks/a.md")];
-    let cases: Vec<(&str, &str)> = vec![
-        ("tasks/a.md", "exact match"),
-        ("./tasks/a.md", "leading ./ normalized"),
-        ("tasks\\a.md", "backslash separator"),
-    ];
-    for (parent, label) in cases {
-        assert_eq!(
-            resolve_parent_for_new_task(parent, &tasks),
-            Some(0),
-            "{label}"
-        );
-    }
-}
-
-#[test]
-fn resolve_parent_for_new_task_miss_cases_with_existing_task() {
-    let tasks = vec![task_without_parent("tasks/a.md")];
-    let cases: Vec<(&str, &str)> = vec![
-        ("tasks/missing.md", "no matching path"),
-        ("", "empty parent string"),
-        ("/abs/path.md", "absolute path"),
-        ("C:\\foo.md", "windows drive prefix"),
-    ];
-    for (parent, label) in cases {
-        assert_eq!(resolve_parent_for_new_task(parent, &tasks), None, "{label}");
-    }
-}
-
-#[test]
-fn resolve_parent_for_new_task_returns_none_for_empty_tasks() {
-    let tasks: Vec<Task> = Vec::new();
-
-    assert_eq!(resolve_parent_for_new_task("tasks/a.md", &tasks), None);
-}
-
-#[test]
-fn validate_chain_from_parent_single_root_returns_ok() {
-    let tasks = vec![task_without_parent("tasks/a.md")];
-
-    assert_eq!(validate_chain_from_parent(0, &tasks), Ok(()));
-}
-
-#[test]
-fn validate_chain_from_parent_edge_19_is_allowed() {
-    let tasks = parent_chain_with_edge_count(19);
-
-    assert_eq!(validate_chain_from_parent(0, &tasks), Ok(()));
-}
-
-#[test]
-fn validate_chain_from_parent_edge_20_returns_too_deep() {
-    let tasks = parent_chain_with_edge_count(20);
-
-    assert_eq!(
-        validate_chain_from_parent(0, &tasks),
-        Err(ParentHierarchyErrorReason::TooDeep),
-    );
-}
-
-#[test]
-fn validate_chain_from_parent_detects_cycle() {
-    let tasks = vec![
-        task_with_parent("tasks/a.md", "tasks/b.md"),
-        task_with_parent("tasks/b.md", "tasks/a.md"),
-    ];
-
-    assert_eq!(
-        validate_chain_from_parent(0, &tasks),
-        Err(ParentHierarchyErrorReason::Cycle),
-    );
-}
-
-#[test]
-fn task_index_build_children_via_aggregate() {
-    let tasks = vec![
-        task_with_parent("tasks/child.md", "tasks/parent.md"),
-        task_without_parent("tasks/parent.md"),
-    ];
-
-    let index = TaskIndex::new(tasks).build_children().unwrap();
-    let result = index.into_tasks();
-    let parent = result
-        .iter()
-        .find(|t| t.file_path == "tasks/parent.md")
-        .unwrap();
-    assert_eq!(parent.children, vec![TaskFilePath::from("tasks/child.md")]);
-}
-
-#[test]
-fn task_index_validate_parent_hierarchy_via_aggregate() {
-    let tasks = vec![
-        task_with_parent("tasks/a.md", "tasks/b.md"),
-        task_with_parent("tasks/b.md", "tasks/a.md"),
-    ];
-
-    let result = TaskIndex::new(tasks).validate_parent_hierarchy();
-    assert!(matches!(result, Err(TaskParseError::CycleOrTooDeep { .. })));
-}
-
-#[test]
-fn task_index_resolve_parent_for_new_task_via_aggregate() {
-    let tasks = vec![task_without_parent("tasks/parent.md")];
-    let index = TaskIndex::new(tasks);
-
-    assert_eq!(
-        index.resolve_parent_for_new_task("tasks/parent.md"),
-        Some(0)
-    );
-    assert_eq!(index.resolve_parent_for_new_task("tasks/missing.md"), None);
-}
-
-#[test]
-fn task_json_byte_level_round_trip() {
-    // VO 化前後で Task JSON 形状が変わらないことを保証する round-trip テスト。
-    // すべての必須フィールド + 1 件の optional (priority) を含む代表的なペイロード。
-    let json = r#"{"id":"tasks/foo.md","filePath":"tasks/foo.md","title":"Fix bug","status":"Doing","priority":"High","labels":["bug","api"],"parent":"tasks/parent.md","links":["tasks/related.md"],"children":["tasks/child.md"],"reverseLinks":["tasks/source.md"],"body":"description","extras":{},"warnings":[]}"#;
-    let parsed: Task = serde_json::from_str(json).unwrap();
-    let serialized = serde_json::to_string(&parsed).unwrap();
-    assert_eq!(serialized, json);
-}
-
-#[test]
-fn task_json_round_trip_omits_none_optional_fields() {
-    // priority / parent が None のケース (skip_serializing_if = Option::is_none)。
-    let json = r#"{"id":"tasks/foo.md","filePath":"tasks/foo.md","title":"Fix bug","status":"Doing","labels":[],"links":[],"children":[],"reverseLinks":[],"body":"","extras":{},"warnings":[]}"#;
-    let parsed: Task = serde_json::from_str(json).unwrap();
-    let serialized = serde_json::to_string(&parsed).unwrap();
-    assert_eq!(serialized, json);
-}
-
-// ============================================================================
-// TaskIndex::insert_new_task_into_cache
-// ============================================================================
-
-fn cache_from(tasks: Vec<Task>) -> HashMap<PathBuf, Task> {
-    let mut cache = HashMap::new();
-    for task in tasks {
-        cache.insert(PathBuf::from(task.file_path.as_str()), task);
-    }
-    cache
-}
-
-fn task_with_links_and_parent(path: &str, parent: Option<&str>, links: &[&str]) -> Task {
-    let mut s = String::from("---\ntitle: Task\nstatus: Todo\n");
-    if let Some(p) = parent {
-        s.push_str(&format!("parent: {p}\n"));
-    }
-    if !links.is_empty() {
-        s.push_str("links:\n");
-        for l in links {
-            s.push_str(&format!("  - {l}\n"));
-        }
-    }
-    s.push_str("---\n");
-    task_from(&s, path)
-}
-
-#[test]
-fn insert_new_task_into_empty_cache_adds_one_entry() {
-    let mut cache = HashMap::new();
-    let new_task = task_without_parent("tasks/new.md");
-
-    let returned = TaskIndex::insert_new_task_into_cache(&mut cache, new_task.clone());
-
-    assert_eq!(1, cache.len());
-    assert!(returned.children.is_empty());
-    assert!(returned.reverse_links.is_empty());
-    assert_eq!(returned.file_path, "tasks/new.md");
-    assert!(cache.contains_key(&PathBuf::from("tasks/new.md")));
-}
-
-#[test]
-fn insert_new_task_appends_to_parent_children_when_parent_exists() {
-    let parent = task_without_parent("tasks/parent.md");
-    let mut cache = cache_from(vec![parent]);
-    let new_task = task_with_parent("tasks/child.md", "tasks/parent.md");
-
-    TaskIndex::insert_new_task_into_cache(&mut cache, new_task);
-
-    let updated_parent = cache.get(&PathBuf::from("tasks/parent.md")).unwrap();
-    assert_eq!(
-        vec![TaskFilePath::from("tasks/child.md")],
-        updated_parent.children
-    );
-}
-
-#[test]
-fn insert_new_task_appends_to_target_reverse_links_when_link_exists() {
-    let target = task_without_parent("tasks/target.md");
-    let mut cache = cache_from(vec![target]);
-    let new_task = task_with_links_and_parent("tasks/source.md", None, &["tasks/target.md"]);
-
-    TaskIndex::insert_new_task_into_cache(&mut cache, new_task);
-
-    let updated_target = cache.get(&PathBuf::from("tasks/target.md")).unwrap();
-    assert_eq!(
-        vec![TaskFilePath::from("tasks/source.md")],
-        updated_target.reverse_links
-    );
-}
-
-#[test]
-fn insert_new_task_resolves_incoming_parent_into_new_task_children() {
-    // 既存 task が new_task を parent として指している（dangling parent）→
-    // 解決して new_task.children に入る。
-    let existing = task_with_parent("tasks/a.md", "tasks/new.md");
-    let mut cache = cache_from(vec![existing]);
-    let new_task = task_without_parent("tasks/new.md");
-
-    let returned = TaskIndex::insert_new_task_into_cache(&mut cache, new_task);
-
-    assert_eq!(vec![TaskFilePath::from("tasks/a.md")], returned.children);
-    // cache 側も同じ children を持つ
-    let cached_new = cache.get(&PathBuf::from("tasks/new.md")).unwrap();
-    assert_eq!(vec![TaskFilePath::from("tasks/a.md")], cached_new.children);
-}
-
-#[test]
-fn insert_new_task_resolves_incoming_links_into_new_task_reverse_links() {
-    let existing = task_with_links_and_parent("tasks/source.md", None, &["tasks/new.md"]);
-    let mut cache = cache_from(vec![existing]);
-    let new_task = task_without_parent("tasks/new.md");
-
-    let returned = TaskIndex::insert_new_task_into_cache(&mut cache, new_task);
-
-    assert_eq!(
-        vec![TaskFilePath::from("tasks/source.md")],
-        returned.reverse_links
-    );
-}
-
-#[test]
-fn insert_new_task_leaves_cache_unchanged_when_parent_and_links_dangling() {
-    let mut cache = HashMap::new();
-    let new_task = task_with_links_and_parent(
-        "tasks/new.md",
-        Some("tasks/missing-parent.md"),
-        &["tasks/missing-link.md"],
-    );
-
-    let returned = TaskIndex::insert_new_task_into_cache(&mut cache, new_task);
-
-    assert_eq!(1, cache.len(), "only new_task inserted");
-    assert!(returned.children.is_empty());
-    assert!(returned.reverse_links.is_empty());
-}
-
-#[test]
-fn insert_new_task_dedups_repeated_link_target_into_single_reverse_link() {
-    let target = task_without_parent("tasks/target.md");
-    let mut cache = cache_from(vec![target]);
-    // links に同じ target が 2 回登場（既存 lenient deserializer により dedupe 済みになるが、
-    // ここでは仕様確認のため両方持つ Task を直接組み立てる）。
-    let mut new_task = task_without_parent("tasks/source.md");
-    new_task.links = vec![
-        TaskFilePath::from("tasks/target.md"),
-        TaskFilePath::from("tasks/target.md"),
-    ];
-
-    TaskIndex::insert_new_task_into_cache(&mut cache, new_task);
-
-    let updated_target = cache.get(&PathBuf::from("tasks/target.md")).unwrap();
-    assert_eq!(
-        vec![TaskFilePath::from("tasks/source.md")],
-        updated_target.reverse_links,
-        "duplicate link target should produce only one reverse_link"
-    );
-}
-
-#[test]
-fn insert_new_task_appends_to_existing_children_at_end_regardless_of_lex_order() {
-    // 順序の不変条件: 差分挿入は既存 parent.children の末尾に新規 task を push する
-    // （`file_path` の辞書順では新規が先頭に来てしまうケースでも変わらない）。
-    let parent = task_without_parent("tasks/zzz-parent.md");
-    // 既存 child a が parent の children に登録済み（pre-resolved 状態を再現）。
-    let mut a = task_with_parent("tasks/m-child.md", "tasks/zzz-parent.md");
-    a.children = Vec::new();
-    let mut parent_pre = parent;
-    parent_pre.children = vec![TaskFilePath::from("tasks/m-child.md")];
-    let mut cache = cache_from(vec![parent_pre, a]);
-
-    // 新規 a-child (辞書順で m-child より前) を追加。
-    let new_task = task_with_parent("tasks/a-child.md", "tasks/zzz-parent.md");
-    TaskIndex::insert_new_task_into_cache(&mut cache, new_task);
-
-    let parent_now = cache.get(&PathBuf::from("tasks/zzz-parent.md")).unwrap();
-    assert_eq!(
-        vec![
-            TaskFilePath::from("tasks/m-child.md"),
-            TaskFilePath::from("tasks/a-child.md"),
-        ],
-        parent_now.children,
-        "new child should be appended at end, not lex-sorted"
-    );
-}
-
-#[test]
-fn insert_new_task_matches_full_rebuild_with_build_children_and_build_reverse_links() {
-    // 差分挿入後の cache snapshot が、全件再構築の結果と等価であることを確認する。
-    // 前提: open_project の挙動と同じく、既存 cache は children/reverse_links を
-    //       既に解決済みの状態から始める。create_task はそこに新規 Task を 1 件追加する。
-    let existing = vec![
-        task_with_links_and_parent("tasks/a.md", Some("tasks/new.md"), &["tasks/new.md"]),
-        task_without_parent("tasks/b.md"),
-        task_with_links_and_parent("tasks/c.md", None, &["tasks/new.md", "tasks/b.md"]),
-    ];
-    let new_task = task_with_links_and_parent("tasks/new.md", Some("tasks/b.md"), &["tasks/c.md"]);
-
-    // 差分挿入経路: 既存 cache は build_children + build_reverse_links 済みから出発
-    let prebuilt = TaskIndex::new(existing.clone())
-        .build_children()
-        .expect("no cycle")
-        .build_reverse_links()
-        .into_tasks();
-    let mut cache = cache_from(prebuilt);
-    TaskIndex::insert_new_task_into_cache(&mut cache, new_task.clone());
-    let mut diff_tasks: Vec<Task> = cache.values().cloned().collect();
-    diff_tasks.sort_by(|a, b| a.file_path.cmp(&b.file_path));
-
-    // 全件再構築経路
-    let mut all = existing.clone();
-    all.push(new_task);
-    let rebuilt = TaskIndex::new(all)
-        .build_children()
-        .expect("no cycle")
-        .build_reverse_links()
-        .into_tasks();
-    let mut full_tasks = rebuilt;
-    full_tasks.sort_by(|a, b| a.file_path.cmp(&b.file_path));
-
-    // `children` / `reverse_links` は集合としての等価性のみを保証する仕様。
-    // - 差分挿入: 既存 parent/target の末尾に新規 task を `push` する（incoming は
-    //   `file_path` 昇順で安定）
-    // - 全件再構築: 入力 task 順に build_children / build_reverse_links が走る
-    //   （walkdir 由来の入力順次第で順序が決まる）
-    // どちらも内部表現で順序自体は spec として固定していない（FE 側で必要なら
-    // 表示順を別途決める）。等価性検証は sort 後比較で行う。
-    assert_eq!(full_tasks.len(), diff_tasks.len());
-    for (a, b) in full_tasks.iter().zip(diff_tasks.iter()) {
-        assert_eq!(a.file_path, b.file_path);
-        let mut a_children = a.children.clone();
-        let mut b_children = b.children.clone();
-        a_children.sort();
-        b_children.sort();
-        assert_eq!(
-            a_children, b_children,
-            "children for {} mismatched (as set)",
-            a.file_path
-        );
-        let mut a_rl = a.reverse_links.clone();
-        let mut b_rl = b.reverse_links.clone();
-        a_rl.sort();
-        b_rl.sort();
-        assert_eq!(
-            a_rl, b_rl,
-            "reverse_links for {} mismatched (as set)",
-            a.file_path
-        );
-    }
-}
